+++ Editor/DecisionTreeNodeView.cs
@@ -1,6 +1,7 @@
 using UnityEditor.Experimental.GraphView;
 using KieranCoppins.DecisionTrees;
 using UnityEngine.UIElements;
+using System.Drawing;
 
 namespace KieranCoppins.DecisionTreesEditor
 {
@@ -43,7 +44,7 @@ namespace KieranCoppins.DecisionTreesEditor
             if (Node is RootNode)
                 return;
 
-            Port port = InstantiatePort(Orientation.Vertical, Direction.Input, Port.Capacity.Single, typeof(bool));
+            Port port = InstantiatePort(Orientation.Vertical, Direction.Input, Port.Capacity.Single, typeof(LogicPortType));
             port.portName = "Entry";
             port.name = "main";
             InputPorts.Add("main", port);
@@ -57,9 +58,11 @@ namespace KieranCoppins.DecisionTreesEditor
                 {
                     foreach (var parameter in constructor.GetParameters())
                     {
-                        port = InstantiatePort(Orientation.Vertical, Direction.Input, Port.Capacity.Single, parameter.ParameterType);
+                        port = InstantiatePort(Orientation.Vertical, Direction.Input, Port.Capacity.Single, parameter.ParameterType.GetGenericArguments()[0]);
                         port.portName = parameter.Name;
                         port.name = parameter.Name;
+                        port.portColor = GetColorForType(parameter.ParameterType);
+                        port.tooltip = parameter.ParameterType.GetGenericArguments()[0].ToString();
                         InputPorts.Add(parameter.Name, port);
                         inputContainer.Add(port);
                     }
@@ -78,14 +81,14 @@ namespace KieranCoppins.DecisionTreesEditor
             }
             else if (Node is Decision)
             {
-                Port port = InstantiatePort(Orientation.Vertical, Direction.Output, Port.Capacity.Single, typeof(bool));
+                Port port = InstantiatePort(Orientation.Vertical, Direction.Output, Port.Capacity.Single, typeof(LogicPortType));
                 port.portName = "TRUE";
                 port.name = "TRUE";
                 OutputPorts.Add("TRUE", port);
                 outputContainer.Add(port);
 
 
-                port = InstantiatePort(Orientation.Vertical, Direction.Output, Port.Capacity.Single, typeof(bool));
+                port = InstantiatePort(Orientation.Vertical, Direction.Output, Port.Capacity.Single, typeof(LogicPortType));
                 port.name = "FALSE";
                 port.portName = "FALSE";
                 OutputPorts.Add("FALSE", port);
@@ -93,7 +96,7 @@ namespace KieranCoppins.DecisionTreesEditor
             }
             else if (Node is RootNode)
             {
-                Port port = InstantiatePort(Orientation.Vertical, Direction.Output, Port.Capacity.Single, typeof(bool));
+                Port port = InstantiatePort(Orientation.Vertical, Direction.Output, Port.Capacity.Single, typeof(LogicPortType));
                 port.name = "main";
                 port.portName = "";
                 OutputPorts.Add("main", port);
@@ -101,4 +104,9 @@ namespace KieranCoppins.DecisionTreesEditor
             }
         }
     }
+}
+
+public class LogicPortType
+{
+    public LogicPortType() { }
 }